+++ src/helpers/messageUtil.js
@@ -1,5 +1,5 @@
 import web3 from 'web3';
-import { ErrorHandler } from '@/helpers';
+import { Toast } from '@/helpers';
 const secp256k1 = require('secp256k1');
 const assert = require('assert');
 const createKeccakHash = require('keccak');
@@ -26,7 +26,7 @@ function toBuffer(v) {
       v = Buffer.from(v.toArray());
     } else {
       const err = new Error('invalid type');
-      ErrorHandler(err, false);
+      Toast.responseHandler(err, false);
     }
   }
   return v;
@@ -105,7 +105,7 @@ function ecrecover(hash, v, r, s) {
   const recovery = v - 27;
   if (recovery !== 0 && recovery !== 1) {
     const e = new Error('Invalid signature v value');
-    ErrorHandler(e, false);
+    Toast.responseHandler(e, false);
   }
   const senderPubKey = secp256k1.recover(hash, signature, recovery);
   return secp256k1.publicKeyConvert(senderPubKey, false).slice(1);