+++ src/layouts/AccessWalletLayout/components/MetamaskModal/MetamaskModal.vue
@@ -98,6 +98,7 @@
 <script>
 import CustomerSupport from '@/components/CustomerSupport';
 import { Web3Wallet } from '@/wallets/software';
+import { Toast } from '@/helpers';
 import Web3 from 'web3';
 import { mapGetters } from 'vuex';
 
@@ -138,7 +139,9 @@ export default {
         try {
           await window.ethereum.enable();
         } catch (e) {
+          Toast.responseHandler(e, Toast.WARN);
           this.web3WalletExists = false;
+          return;
         }
         this.signIn(window.web3);
       } else if (window.web3) {
@@ -157,7 +160,8 @@ export default {
             path: 'interface'
           });
         })
-        .catch(() => {
+        .catch(e => {
+          Toast.responseHandler(e, Toast.ERROR);
           return (this.web3WalletExists = false);
         });
     },